Extracted constants from interpreter.h. Extracted commands from liquid.cpp
